move combat_manager to simulation
