+++ src/models/scenes/combat_scene.py
@@ -9,7 +9,7 @@ from models.characters.character_examples import CharacterTypes
 from models.characters.character_impl import build_character
 from models.characters.conditions import IsDead
 from models.characters.player import get_player
-from models.combat.combat_manager_base import CombatManager, valid_moves
+from simulation.combat_manager_base import CombatManager, valid_moves
 from models.combat.moves_base import Move
 from models.scenes import scene_examples
 from models.scenes.scenes_base import Resolution, Scene
