+++ src/simulation/combat_gym.py
@@ -4,8 +4,8 @@ from typing import Any, List, Sequence, Tuple
 
 from models.characters.character_base import Character
 from models.characters.states import Attributes, State
-from models.combat.combat_manager_base import (CombatGroup, CombatManager,
-                                               GroupMove)
+from simulation.combat_manager_base import (CombatGroup, CombatManager,
+                                            GroupMove)
 
 RewardFunc = Any
 CombatState = Sequence[int]
